fixed several potential bugs (thanks to findbugs)

+++ trunk/core/src/test/java/org/codehaus/waffle/bind/OgnlTypeConverterTest.java
@@ -34,7 +34,7 @@ public class OgnlTypeConverterTest {
         OgnlTypeConverter converter = new OgnlTypeConverter();
         Integer value = (Integer) converter.convertValue("foobar", "15", Integer.class);
 
-        assertEquals(15, value.intValue());
+        assertEquals(15, value);
     }
 
     @Test
